+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## Development ()
+
+* **Update**: Updated TEI editor to address duplicate metadata issue
+
 ## 0.8.8 (02.12.2019)
 
 * **Bugfix**: Missing menu options in the editor